+++ apps/web/app/(auth)/auth/verify/page.tsx
@@ -2,7 +2,7 @@
 
 import { SignIn } from "@/components/auth/SignIn";
 import { useSearchParams } from "next/navigation";
-import FormWrapper from "@/components/auth/FormWrapper";
+import FormWrapper from "@/app/(auth)/auth/FormWrapper";
 
 export default function Verify() {
   const searchParams = useSearchParams();
